Update CSS and JS defaults.
Introduce Just as a dependency for utilities. See https://github.com/angus-c/just. They are dependency free and smaller than using lodash.
Remove utilities that are not often used. These can be added as-and-when. In the case of 'findParentWith...', this can now be replaced with the default 'closest()' method.
Remove KeyboardFocus, this can be added in as-and-when. Fixes #1